requirements.txt
langchain==0.3.7
langchain-community==0.3.5
langchain-openai==0.2.6
langchain-chroma==0.1.4
chromadb==0.5.18
openai==1.54.4
streamlit==1.40.1
fastapi==0.115.5
uvicorn==0.32.0
python-dotenv==1.0.1
pypdf==5.1.0
beautifulsoup4==4.12.3
requests==2.32.3
tiktoken==0.8.0
rank-bm25==0.2.2
pydantic==2.10.0
